+++ packages/core/src/applicability/index.ts
@@ -1,2 +1,3 @@
+export * from "./invariant";
 export * from "./resolver";
 export * from "./types";